moved utils function to director-v2
