+++ packages/models-library/src/models_library/clusters.py
@@ -3,7 +3,7 @@ from typing import Any, Dict, Literal, Optional, Union
 
 from pydantic import AnyUrl, BaseModel, Extra, Field, HttpUrl, SecretStr, validator
 from pydantic.types import NonNegativeInt
-from settings_library.utils_cli import create_json_encoder_wo_secrets
+
 from simcore_postgres_database.models.clusters import ClusterType
 
 from .users import GroupID
@@ -106,18 +106,6 @@ class BaseCluster(BaseModel):
     class Config:
         extra = Extra.forbid
         use_enum_values = True
-
-    def to_clusters_db(self, only_update: bool) -> Dict[str, Any]:
-        db_model = json.loads(
-            self.json(
-                by_alias=True,
-                exclude={"id", "access_rights"},
-                exclude_unset=only_update,
-                exclude_none=only_update,
-                encoder=create_json_encoder_wo_secrets(BaseCluster),
-            )
-        )
-        return db_model
 
 
 ClusterID = NonNegativeInt
